Extend workers and services to support namespace for Security Policies
Changelog: added
EE: true

+++ ee/app/workers/security/create_orchestration_policy_worker.rb
@@ -26,12 +26,14 @@ module Security
               .execute
           end
 
-          configuration.transaction do
-            configuration.approval_rules.scan_finding.delete_all
-            configuration.active_scan_result_policies.each_with_index do |policy, policy_index|
-              Security::SecurityOrchestrationPolicies::ProcessScanResultPolicyService
-                .new(policy_configuration: configuration, policy: policy, policy_index: policy_index)
-                .execute
+          if configuration.project?
+            configuration.transaction do
+              configuration.approval_rules.scan_finding.delete_all
+              configuration.active_scan_result_policies.each_with_index do |policy, policy_index|
+                Security::SecurityOrchestrationPolicies::ProcessScanResultPolicyService
+                  .new(policy_configuration: configuration, policy: policy, policy_index: policy_index)
+                  .execute
+              end
             end
           end
